feat : 공고 상세페이지 ( 사장님, 일반회원)
